feat: use multikey
This addresses usage of `publicKeyMultibase` that was out of spec with
what was intended in the DID Core spec. Namely, `publicKeyMultibase`
only expresses the encoding of the value and does not necessarily
include the multicodec binary header. Presence or absence of this header
should be determined by the verification method type definition.
[Multikey](https://www.w3.org/TR/vc-data-integrity/#multikey)
accomplishes this goal and also cleans up some aspects of the spec and
this implementation.

Signed-off-by: Daniel Bluhm <dbluhm@pm.me>